Go to end of chain before calling .then
This version accomodates overwritten .then's, but breaks a few Promises/A+ tests.

It strategy for the scenario at https://github.com/promises-aplus/promises-spec/issues/179#issuecomment-93878533 is:

- Add a [[PromiseFollowing]] slot to all promises
- Resolving x1 to x0 notices that x0 is a promise without any value for its [[PromiseFollowing]] internal slot, so:
  - Do the normal thing, i.e. `x0.then(resolveX1, rejectX1)`
  - Save x1@[[PromiseFollowing]] = x0
- Resolving x2 to x1 notices that x1 is a promise with a value for its [[PromiseFollowing]] internal slot, so:
  - Follow the chain of [[PromiseFollowing]] to get a "real" resolution value of x0
  - Now do as before, i.e. `x0.then(resolveX2, rejectX2)` and x2@[[PromiseFollowing]] = x0.

Thus the reference graph is ... fuck, it's just x0 -> x1, x2 so x1 stays in memory.

This breaks some Promises/A+ tests which are coded in terms of a given thenable only having its `then` property accessed once.